chore: moved utility functions to utils directory
